[Nexters/covid-letter-web#83] :lipstick: CSS 수정
