+++ src/pages/covid/letter/new/[optionId].tsx
@@ -36,11 +36,10 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 
 const Container = styled.section`
     ${tw`tw-bg-beige-300 tw-flex tw-flex-col`}
-    //max-width: 36rem;
-    height: calc(100vh - 5.2rem);
     margin: 0 auto;
     text-align: center;
     align-items: center;
+    //min-height: calc(100vh - 5.2rem);
 `
 const Header = styled.section`
     height: 5.6rem;
